fix map 100% and start fix statistics

+++ GameEngine/Profile.php
@@ -54,7 +54,9 @@ class Profile {
 		$database->submitProfile($database->RemoveXSS($post['uid']),$database->RemoveXSS($post['mw']),$database->RemoveXSS($post['ort']),$database->RemoveXSS($birthday),$database->RemoveXSS($post['be2']),$database->RemoveXSS($post['be1']));
 		$varray = $database->getProfileVillages($post['uid']);
 		for($i=0;$i<=count($varray)-1;$i++) {
+			if(trim($post['dname'.$i]) != ""){
 			$database->setVillageName($database->RemoveXSS($varray[$i]['wref']),$database->RemoveXSS($post['dname'.$i]));
+			}
 		}
 		header("Location: ?uid=".$post['uid']);
 	}
